did something

+++ LaserAge/EasyEnemy.h
@@ -13,7 +13,7 @@ public:
 	
 	
 	EasyEnemy(String Name, float X, float Y, int W, int H);
-	void update(float time, int width, int heigth);
+	void update(int time, int width, int heigth);
 	
 };
