+++ LaserAge/HardEnemy.h
@@ -10,7 +10,7 @@ class HardEnemy :public Enemy
 public:
 	
 	HardEnemy(String Name, float X, float Y, int W, int H);
-	void update(float time, int width, int heigth);
+	void update(int time, int width, int heigth);
 
 };
