+++ LaserAge/NormalEnemy.h
@@ -14,6 +14,6 @@ public:
 	
 	NormalEnemy(String Name, float X, float Y, int W, int H);
 
-	void update(float time, int width, int heigth);
+	void update(int time, int width, int heigth);
 };
